+++ libraries/Menu/examples/softboard/menu_IO_configuration.h
@@ -16,7 +16,7 @@
 #ifndef MENU_CONFIGURATION_H
 
 #if ! defined EOF32
-  #define EOF32		0xffffffff
+  #define EOF32		(int) 0xffffffff
 #endif
 
 /*
